requirements.txt
transformers==4.38.2
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
frozendict==2.3.8
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.34
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.4
openai==0.28
outcome==1.2.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
plotly==5.16.1
protobuf==4.24.2
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.18.0
selenium==4.12.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.28
zipp==3.16.2
tensorflow==2.16.1
tf-keras==2.16.0